scrapers/requirements.txt
scrapy==2.11.1
scrapy-playwright==0.0.40
playwright==1.44.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
itemadapter==0.9.0
w3lib==2.1.2
fake-useragent==1.5.1
tenacity==8.3.0          # Retry logic
loguru==0.7.2            # Logging estructurado
pydantic==2.7.1          # Validación de items
